move remaining actions from app.rs into new file app_actions.rs
